+++ src/ios/CameraSessionManager.h
@@ -8,6 +8,8 @@
 - (void) setupSession:(NSString *)defaultCamera completion:(void(^)(BOOL started))completion options:(NSDictionary *)options photoSettings:(AVCapturePhotoSettings *)photoSettings;
 - (void) setFlashMode:(NSInteger)flashMode photoSettings:(AVCapturePhotoSettings *)photoSettings;
 - (void) torchSwitch:(NSInteger)torchState;
+- (void) switchCameraTo:(NSString *)cameraMode completion:(void (^)(BOOL success))completion;
+- (BOOL) deviceHasUltraWideCamera;
 - (void) updateOrientation:(AVCaptureVideoOrientation)orientation;
 - (AVCaptureVideoOrientation) getCurrentOrientation:(UIInterfaceOrientation)toInterfaceOrientation;
 + (AVCaptureSessionPreset) calculateResolution:(NSInteger)targetSize;